==> P1 - get_sec_data.ipynb <==
# P1 - Download Financial Data for S&P500 Companies from SEC

## Overview:

This project aims to download financial data for S&P500 companies from SEC filings using the SEC API. The data gathered will be applied in the subsequent project to compute key financial metrics for factor-based investment analysis.

Techniques used in the project:

1. Utilizing web scraping and API requests to collect company tickers and financial data. This includes interacting with the SEC database via submission and company fact APIs.
2. Processing JSON files and leveraging the requests library for API retrieval.
3. Manipulating time series data by resampling and aligning the datasets over uniform time periods.
4. Cleaning and adjusting the data, particularly dealing with cumulative figures and converting them into quarterly amounts for consistency.

# import libraries
import yfinance as yf
import pandas as pd
import requests
import json
import numpy as np
import pickle
import copy
import datetime as dt
import os

headers = {"User-Agent": "ian.ye.fu@gmail.com"} 

data_folder_download = './datasets/download/'
data_folder_generate = './datasets/generate/'

%store -r removed_tickers_list

## Step 1: Download S&P500 tickers from wikipedia 

# Wikipedia URL for S&P 500
url = 'https://en.wikipedia.org/wiki/List_of_S%26P_500_companies'

# Read the HTML tables from the webpage
tables = pd.read_html(url)

# Table 1: List of S&P 500 companies
sp500_table = tables[0]

# Table 2: Recent changes to the S&P 500
recent_changes_table = tables[1]

# Save these tables to CSV files:
sp500_table.to_csv(data_folder_download + 'sp500_companies.csv', index=False)
recent_changes_table.to_csv(data_folder_download + 'sp500_recent_changes.csv', index=False)

# create a list of S&P500 tickers
sp500_df = pd.read_csv(data_folder_download + 'sp500_companies.csv', index_col = 0)
sp500_tickers = sp500_df.index.to_list()

%store sp500_tickers

sp500_df.head()

## Step 2: Convert the CIK code format to match SEC Fillings

# Create the CIK column and get sp500_cik_list
sp500_cik_list = [str(cik).zfill(10) for cik in sp500_df['CIK'].tolist()]  
sp500_cik = pd.Series(sp500_cik_list, index = sp500_df.index)
sp500_df['CIK'] = sp500_cik

sp500_df.drop(columns = 'Headquarters Location', inplace = True)

sp500_df.head()

sp500_df.loc['AAPL']

with open(data_folder_generate + 'sp500_df_v1.0.pkl', 'wb') as f: 
    pickle.dump(sp500_df, f)

## Step 4: Get all the company_fact data from downloaded archive json files. 

def get_facts(ticker):
    """
    Load company_facts data from local json_file
    """
    cik = f'CIK{sp500_df.loc[ticker, 'CIK']}'
    directory = data_folder_download + 'companyfacts/'
    file_path = os.path.join(directory, f"{cik}.json")  # Example: specific file

    # Load the JSON data
    with open(file_path, 'r') as json_file:
        company_facts = json.load(json_file)

    # Print the keys in the root of the JSON to understand its structure
    return company_facts

# test get_facts function: success.
company_facts_appl = get_facts('AAPL')

## Step 5: Create financials dataframes from the company_facts dict. 

def facts_DF(ticker, headers=headers):
    """
    Convert the company_facts dict to dataframes
    """
    facts = get_facts(ticker)
    us_gaap_data = facts["facts"]["us-gaap"]
    df_data = []
    for fact, details in us_gaap_data.items():
        for unit in details["units"].keys():
            for item in details["units"][unit]:
                row = item.copy() # keep the original data intact
                row["fact"] = fact
                row['label'] = details['label']
                df_data.append(row)

    df = pd.DataFrame(df_data)
    df["end"] = pd.to_datetime(df["end"])
    df["start"] = pd.to_datetime(df["start"])
    df = df.drop_duplicates(subset=["fact", "end", "val"])
    df.set_index("end", inplace=True)
    labels_dict = {fact: details["label"] for fact, details in us_gaap_data.items()}
   
    return df

# test get_facts function: success.
ice_facts_df = facts_DF('ICE', headers)

## Step 6: Define the financial data to be downloaded

pd.set_option('display.max_rows', None)

# get the data labels in the SEC database for all the essential financials for the factor calculation. 
data_category = ['Assets', 
                 'Liabilities', 
                 'LiabilitiesCurrent', 
                 'LiabilitiesNoncurrent', 
                 'LiabilitiesAndStockholdersEquity',  
                 'StockholdersEquity', 
                 'StockholdersEquityIncludingPortionAttributableToNoncontrollingInterest',
                 'EarningsPerShareDiluted', 
                 'CommonStockDividendsPerShareDeclared', 
                 'WeightedAverageNumberOfDilutedSharesOutstanding', 
                 'WeightedAverageNumberOfSharesOutstandingBasic',
                 'CommonStockSharesOutstanding',
                 'NetIncomeLoss' 
                 ]
%store data_category

## Step 7: Download the financial data from SEC 

def download_financial_data_from_SEC(sp500_tickers, data_category):
    """
    download all the financial metrics in the data_category from SEC for the sp500 companies 
    """
    sp500_financial_data = {} # outer dict
    
    for ticker in sp500_tickers:
        
        df = facts_DF(ticker)  # Correct data extraction for the ticker
        financial_data= {} # inner dict
        
        # Loop over all categories in the data_category list
        for category in data_category:
            
            x = df.query('fact == @category')
            # remove the duplicated rows based on 'end' index and keep the last record
            x = x[(x['val']!= 0) & (x['val'].notna())]
            cleaned_data = x[~x.index.duplicated(keep='last')].sort_index(ascending = True)
            # slice data from only 2013 onwards
            financial_data[category] = cleaned_data.loc['2013':]  
            
        # Assign the financial data for each ticker
        sp500_financial_data[ticker] = financial_data
        
    return  sp500_financial_data

sp500_financial_data = download_financial_data_from_SEC(sp500_tickers, data_category)

with open(data_folder_generate + 'sp500_financial_data_v1.0.pkl', 'wb') as f: 
    pickle.dump(sp500_financial_data, f)

## Step 8: Clean the financial data from SEC 

sp500_financial_data_updated = convert_annual_to_quarter(sp500_financial_data, updated_data_category)

# notice only I/S based metrics need to be updated.
updated_data_category = [
                 'EarningsPerShareDiluted',
                 'CommonStockDividendsPerShareDeclared', 
                 'NetIncomeLoss'
                 ]

def convert_annual_to_quarter(sp500_financial_data, updated_data_category):
    """
    This function updates financial data by subtracting the annual rows by 
    the sum of the previous three quarters for a given list of tickers and data categories
    """
    
    for ticker in sp500_financial_data.keys(): 

        for category in sp500_financial_data[ticker].keys(): 
            
            new_df = sp500_financial_data[ticker][category].reset_index().copy() 

            if category == "EarningsPerShareDiluted" or category == 'CommonStockDividendsPerShareDeclared' or category == 'NetIncomeLoss':

                # identify the annual rows
                index_list = new_df[(new_df['end'] - new_df['start']).dt.days > 130].index.tolist()
    
                # subtract the annual rows by the sum of the previous three quarters.
                for i in index_list: 
                    new_df.loc[i,'val'] = new_df.loc[i,'val'] - new_df.loc[i-3: i-1, 'val'].sum()
    
            sp500_financial_data[ticker][category] = new_df.set_index('end')
        
    return sp500_financial_data

sp500_financial_data_updated['ABNB']['StockholdersEquity']

with open(data_folder_generate +'sp500_financial_data_v1.1.pkl', 'wb') as f:
    pickle.dump(sp500_financial_data_updated, f)

## Step 9: Download S&P500 stock daily price data from Yahoo Finance   

syear = 2013
smonth = 1
sday = 1
eyear = 2024
emonth = 9
eday = 1

stocks_not_downloaded = []

folder = '../datasets/download/companyprice/'

interval = '1mo'
def save_to_csv_from_yahoo(folder, ticker, syear, smonth, sday, eyear, emonth, eday):
    """
    Download stock.['Adj Close'] data from Yahoo Finance and save them to folder.
    """
    start = dt.datetime(syear, smonth, sday)
    end = dt.datetime(eyear, emonth, eday)
    print('Get Data for:', ticker)
    df = yf.download(ticker, start, end)['Adj Close']
    
    if df.empty:
        print("Couldn't Get Data for: ", ticker)
        stocks_not_downloaded.append(ticker)
    df.to_csv(folder + ticker + '.csv')

# convert the '.' ticker format to '-' ticker format.
tickers_not_downloaded = [ticker.replace('.', '-') for ticker in stocks_not_downloaded]

for ticker in sp500_df.index.to_list(): 
    save_to_csv_from_yahoo(folder, ticker, syear, smonth, sday, eyear, emonth, eday)

# redownloaded the failed tickers 
for ticker in tickers_not_downloaded: 
    save_to_csv_from_yahoo(folder, ticker, syear, smonth, sday, eyear, emonth, eday, interval)

for ticker in removed_tickers_list: 
    save_to_csv_from_yahoo(folder, ticker, syear, smonth, sday, eyear, emonth, eday)

# download cap-weighted sp500 index historical performance. 
save_to_csv_from_yahoo('./', '^GSPC', syear, smonth, sday, eyear, emonth, eday)

# download equal-weighted sp500 index historical performance. 
save_to_csv_from_yahoo('./', '^SPXEW', syear, smonth, sday, eyear, emonth, eday)
