==> P2 - add_removed_tickers.ipynb <==
# Adding Removed S&P 500 Tickers

## Overview:

This step aims to add historically removed S&P500 company tickers to eliminate the survivorship biase among the financial data. 


# import libraries
import yfinance as yf
import pandas as pd
import requests
import json
import numpy as np
import pickle
import copy
import datetime as dt
import os

headers = {"User-Agent": "ian.ye.fu@gmail.com"} 
data_folder_download = './datasets/download/'
data_folder_generate = './datasets/generate/'

## Step 1: Get removed and added tickers

recent_changes_tickers = pd.read_csv(data_folder_download + 'sp500_recent_changes.csv', header = 0, index_col = 'Date')

recent_changes_tickers.info()

# Check the overlap between tickers added and tickers removed
set_added_tickers = set(recent_changes_tickers["Added"].tolist())
set_removed_tickers = set(recent_changes_tickers["Removed"].tolist())
print(f'number of ticker overlaps between removed tickers and added tickers: {len(list(set_added_tickers.intersection(set_removed_tickers)))}')

# create added_tickers_df
added_tickers_df = recent_changes_tickers.iloc[1: ][['Added', 'Added.1']].dropna(axis = 0)
added_tickers_df.index = pd.to_datetime(added_tickers_df.index, format='%B %d, %Y')
added_tickers_df = added_tickers_df.reset_index().rename(columns = {'Date':'Date_Added', 'Added': 'Added_Ticker', 'Added.1': 'Security_Added'}).set_index('Added_Ticker')

#Only get the tickers added after 2013 for the analysis period between 2013 to 2024.
added_tickers_df = added_tickers_df.query('Date_Added >= "2013"')
with open(data_folder_generate + 'added_tickers_df.pkl', 'wb') as f: 
    pickle.dump(added_tickers_df, f)

with open(data_folder_generate + 'added_tickers_df.pkl', 'rb') as f: 
    added_tickers_df = pickle.load(f)

added_tickers_df.head()

# create removed_tickers_df
removed_tickers_df = recent_changes_tickers.iloc[1: ][['Removed', 'Removed.1']].dropna(axis = 0)
removed_tickers_df.index = pd.to_datetime(removed_tickers_df.index, format='%B %d, %Y')
removed_tickers_df = removed_tickers_df.reset_index().rename(columns = {'Date':'Date_Removed', 'Removed': 'Removed_Ticker', 'Removed.1': 'Security_Removed'}).set_index('Removed_Ticker')

removed_tickers_df = removed_tickers_df.query('Date_Removed >= "2013"')
with open(data_folder_generate + 'removed_tickers_df.pkl', 'wb') as f: 
    pickle.dump(removed_tickers_df, f)

with open(data_folder_generate + 'removed_tickers_df.pkl', 'rb') as f: 
    removed_tickers_df = pickle.load(f)

removed_tickers_df.head()

## Step 2: Get the CIK numbers for removed tickers

def cik_matching_ticker(ticker, headers=headers):
    """
    get the cik number for tickers
    """
    ticker = ticker.upper().replace(".", "-")
    ticker_json = requests.get(
        "https://www.sec.gov/files/company_tickers.json", headers=headers
    ).json()

    for company in ticker_json.values():
        if company["ticker"] == ticker:
            cik = str(company["cik_str"]).zfill(10)
            return cik
    raise ValueError(f"Ticker {ticker} not found in SEC database")

removed_tickers_list = removed_tickers_df.index.tolist()

for ticker in removed_tickers_list:
    try:
        print(f"Fetching CIK for {ticker}")  
        removed_tickers_df.loc[ticker, 'CIK'] = cik_matching_ticker(ticker, headers)
    except ValueError as e:
        print(f"Error for ticker {ticker}: {e}")
        continue  

removed_tickers_df = removed_tickers_df.rename(columns = {'CIK':'removed_cik'})
removed_tickers_df

with open(data_folder_generate + 'sp500_df.pkl', 'rb') as f: 
    sp500_df = pickle.load(f)

sp500_df.head()

## Step 3. Add the removed tickers to the sp500 tickers table

# merge the sp500_df with added_date
merged_df_added_date = pd.merge(left = sp500_df, right = added_tickers_df, left_index = True, right_index = True, how = 'outer')

#merge updated sp500_df with removed_date
merged_df_both_dates = pd.merge(left = merged_df_added_date, right = removed_tickers_df, left_index = True, right_index = True, how = 'outer')

merged_df_both_dates = merged_df_both_dates.reset_index()

merged_df_both_dates.rename(columns = {'Symbol':'Ticker'}, inplace = True)

merged_df_both_dates

## Step 4. Clean up the merged sp500 tickers table

# Fill 'Security' with values from 'Security Added' and 'Security Removed'
merged_df_both_dates['Security'] = merged_df_both_dates['Security'].fillna(merged_df_both_dates['Security_Added']).fillna(merged_df_both_dates['Security_Removed'])

# Fill 'sp500cik' with values from 'removed_cik'
merged_df_both_dates['CIK'] = merged_df_both_dates['CIK'].fillna(merged_df_both_dates['removed_cik'])

# Fill 'Date added' with values from 'Date_Added'
merged_df_both_dates['Date added'] = merged_df_both_dates['Date added'].fillna(merged_df_both_dates['Date_Added'])

updated_merged_df_both_dates = merged_df_both_dates[~merged_df_both_dates['CIK'].isna()].copy()

updated_merged_df_both_dates = updated_merged_df_both_dates.drop(columns = ['Date_Added', 'Security_Added', 'Security_Removed', 'removed_cik'])

# filter out the tickers that don't have cik code on SEC.
updated_merged_df_both_dates_df = updated_merged_df_both_dates.loc[updated_merged_df_both_dates['CIK'] != "nan"].copy()

# Add the GICS sector label to the newly added tickers
updated_merged_df_both_dates_df['GICS Sector'].fillna(updated_merged_df_both_dates_df['Ticker'].map(removed_tickers_GICS_sector_category), inplace=True)

updated_merged_df_both_dates_df = updated_merged_df_both_dates_df.rename(columns = {'Date added':'Date_Added'})
new_column_order = ['Ticker', 'Security', 'GICS Sector', 'GICS Sub-Industry', 'Date_Added', 'Date_Removed', 'Founded', 'CIK']
reindexed_merged_sp500_df = updated_merged_df_both_dates_df[new_column_order].set_index('Ticker')
reindexed_merged_sp500_df

reindexed_merged_sp500_df = reindexed_merged_sp500_df.query('Security != "Under Armour"')

with open(data_folder_generate + 'sp500_df_v2.0.pkl', 'wb') as f:
    pickle.dump(reindexed_merged_sp500_df, f)

removed_tickers_GICS_sector_category = {
    'AA': 'Materials',
    'AAP': 'Consumer Discretionary',
    'ABMD': 'Health Care',
    'ACE': 'Financials',
    'ADS': 'Information Technology',
    'ADT': 'Industrials',
    'AET': 'Health Care',
    'AGN': 'Health Care',
    'AIV': 'Real Estate',
    'ALK': 'Industrials',
    'ALTR': 'Information Technology',
    'ALXN': 'Health Care',
    'AMG': 'Financials',
    'AN': 'Consumer Discretionary',
    'ANDV': 'Energy',
    'ANF': 'Consumer Discretionary',
    'APC': 'Energy',
    'APOL': 'Consumer Discretionary',
    'ARG': 'Materials',
    'ARNC': 'Materials',
    'ATI': 'Materials',
    'ATVI': 'Communication Services',
    'AVP': 'Consumer Staples',
    'AYI': 'Industrials',
    'BBBY': 'Consumer Discretionary',
    'BCR': 'Health Care',
    'BEAM': 'Health Care',
    'BHF': 'Financials',
    'BIG': 'Consumer Discretionary',
    'BMC': 'Information Technology',
    'BMS': 'Materials',
    'BRCM': 'Information Technology',
    'BTU': 'Energy',
    'BXLT': 'Health Care',
    'CA': 'Information Technology',
    'CAM': 'Energy',
    'CCE': 'Consumer Staples',
    'CDAY': 'Information Technology',
    'CELG': 'Health Care',
    'CERN': 'Health Care',
    'CFN': 'Health Care',
    'CHK': 'Energy',
    'CLF': 'Materials',
    'CMA': 'Financials',
    'CMCSK': 'Communication Services',
    'CNX': 'Energy',
    'COL': 'Industrials',
    'COTY': 'Consumer Staples',
    'COV': 'Health Care',
    'CPGX': 'Energy',
    'CPRI': 'Consumer Discretionary',
    'CSC': 'Information Technology',
    'CSRA': 'Information Technology',
    'CTXS': 'Information Technology',
    'CVC': 'Communication Services',
    'CVH': 'Health Care',
    'CXO': 'Energy',
    'DELL': 'Information Technology',
    'DF': 'Consumer Staples',
    'DISCK': 'Communication Services',
    'DISH': 'Communication Services',
    'DNB': 'Information Technology',
    'DNR': 'Energy',
    'DO': 'Energy',
    'DPS': 'Consumer Staples',
    'DRE': 'Real Estate',
    'DTV': 'Communication Services',
    'DWDP': 'Materials',
    'DXC': 'Information Technology',
    'EMC': 'Information Technology',
    'ENDP': 'Health Care',
    'ESRX': 'Health Care',
    'ESV': 'Energy',
    'ETFC': 'Financials',
    'EVHC': 'Health Care',
    'FB': 'Communication Services',
    'FBHS': 'Consumer Discretionary',
    'FDO': 'Consumer Discretionary',
    'FHN': 'Financials',
    'FII': 'Financials',
    'FL': 'Consumer Discretionary',
    'FLIR': 'Information Technology',
    'FLR': 'Industrials',
    'FLS': 'Industrials',
    'FLT': 'Information Technology',
    'FOSL': 'Consumer Discretionary',
    'FRC': 'Financials',
    'FRX': 'Health Care',
    'FTI': 'Energy',
    'FTR': 'Communication Services',
    'GAS': 'Utilities',
    'GGP': 'Real Estate',
    'GHC': 'Consumer Discretionary',
    'GMCR': 'Consumer Staples',
    'GME': 'Consumer Discretionary',
    'GNW': 'Financials',
    'GPS': 'Consumer Discretionary',
    'GT': 'Consumer Discretionary',
    'HAR': 'Consumer Discretionary',
    'HBI': 'Consumer Discretionary',
    'HCBK': 'Financials',
    'HFC': 'Energy',
    'HNZ': 'Consumer Staples',
    'HOG': 'Consumer Discretionary',
    'HOT': 'Consumer Discretionary',
    'HP': 'Energy',
    'HRB': 'Financials',
    'HSP': 'Health Care',
    'IGT': 'Consumer Discretionary',
    'ILMN': 'Health Care',
    'INFO': 'Information Technology',
    'IPGP': 'Information Technology',
    'JCP': 'Consumer Discretionary',
    'JDSU': 'Information Technology',
    'JEF': 'Financials',
    'JOY': 'Industrials',
    'JWN': 'Consumer Discretionary',
    'KORS': 'Consumer Discretionary',
    'KRFT': 'Consumer Staples',
    'KSS': 'Consumer Discretionary',
    'KSU': 'Industrials',
    'LEG': 'Consumer Discretionary',
    'LIFE': 'Health Care',
    'LLL': 'Industrials',
    'LLTC': 'Information Technology',
    'LM': 'Financials',
    'LNC': 'Financials',
    'LO': 'Consumer Staples',
    'LSI': 'Real Estate',
    'LUMN': 'Information Technology',
    'LVLT': 'Information Technology',
    'M': 'Consumer Discretionary',
    'MAC': 'Real Estate',
    'MAT': 'Consumer Discretionary',
    'MBC': 'Health Care',
    'MJN': 'Consumer Staples',
    'MNK': 'Health Care',
    'MOLX': 'Information Technology',
    'MON': 'Materials',
    'MUR': 'Energy',
    'MXIM': 'Information Technology',
    'NAVI': 'Financials',
    'NBL': 'Energy',
    'NBR': 'Energy',
    'NE': 'Energy',
    'NFX': 'Energy',
    'NKTR': 'Health Care',
    'NLSN': 'Communication Services',
    'NOV': 'Energy',
    'NWL': 'Consumer Discretionary',
    'NYX': 'Financials',
    'OGN': 'Health Care',
    'OI': 'Materials',
    'PBCT': 'Financials',
    'PBI': 'Industrials',
    'PCL': 'Real Estate',
    'PCP': 'Industrials',
    'PCS': 'Communication Services',
    'PDCO': 'Health Care',
    'PENN': 'Consumer Discretionary',
    'PETM': 'Consumer Discretionary',
    'PLL': 'Materials',
    'POM': 'Utilities',
    'PRGO': 'Health Care',
    'PVH': 'Consumer Discretionary',
    'PXD': 'Energy',
    'Q': 'Communication Services',
    'QEP': 'Energy',
    'R': 'Industrials',
    'RAI': 'Consumer Staples',
    'RDC': 'Energy',
    'RE': 'Financials',
    'RHI': 'Industrials',
    'RHT': 'Information Technology',
    'RIG': 'Energy',
    'RRC': 'Energy',
    'RTN': 'Industrials',
    'S': 'Communication Services',
    'SAI': 'Information Technology',
    'SBNY': 'Financials',
    'SCG': 'Utilities',
    'SE': 'Communication Services',
    'SEDG': 'Information Technology',
    'SEE': 'Materials',
    'SIAL': 'Materials',
    'SIG': 'Consumer Discretionary',
    'SIVB': 'Financials',
    'SLG': 'Real Estate',
    'SLM': 'Financials',
    'SNDK': 'Information Technology',
    'SNI': 'Communication Services',
    'SPLS': 'Consumer Discretionary',
    'SRCL': 'Industrials',
    'STI': 'Financials',
    'STJ': 'Health Care',
    'SWN': 'Energy',
    'SWY': 'Consumer Staples',
    'TDC': 'Information Technology',
    'TE': 'Utilities',
    'TEG': 'Utilities',
    'TGNA': 'Communication Services',
    'THC': 'Health Care',
    'TIF': 'Consumer Discretionary',
    'TRIP': 'Communication Services',
    'TSS': 'Information Technology',
    'TWC': 'Communication Services',
    'TWTR': 'Communication Services',
    'TWX': 'Communication Services',
    'TYC': 'Industrials',
    'UA': 'Consumer Discretionary',
    'UAA': 'Consumer Discretionary',
    'UNM': 'Financials',
    'URBN': 'Consumer Discretionary',
    'VAR': 'Health Care',
    'VFC': 'Consumer Discretionary',
    'VIAB': 'Communication Services',
    'VNO': 'Real Estate',
    'VNT': 'Industrials',
    'WCG': 'Health Care',
    'WFM': 'Consumer Staples',
    'WHR': 'Consumer Discretionary',
    'WIN': 'Communication Services',
    'WLTW': 'Financials',
    'WPX': 'Energy',
    'WU': 'Information Technology',
    'WYN': 'Consumer Discretionary',
    'X': 'Materials',
    'XEC': 'Energy',
    'XL': 'Financials',
    'XLNX': 'Information Technology',
    'XRAY': 'Health Care',
    'XRX': 'Information Technology',
    'YHOO': 'Communication Services',
    'ZION': 'Financials'
}
