==> P3 - get_sec_data_for_removed_tickers.ipynb <==
# P2 - Download Removed Tickers from SEC

## Overview:

This project aims to download financial data for historically removed S&P500 companies from SEC filings using the SEC API. This step is necessary to remove the survivorship bias among financial data.

# import libraries
import yfinance as yf
import pandas as pd
import requests
import json
import numpy as np
import pickle
import copy
import datetime as dt
import os
import scripts as s

headers = {"User-Agent": "ian.ye.fu@gmail.com"} 

data_folder_download = '../datasets/download/'
data_folder_generate = '../datasets/generate/'

%store -r data_category

with open(data_folder_generate + 'sp500_df_v1.0.pkl', 'rb') as f:
    current_sp500_df = pickle.load(f)

with open(data_folder_generate + 'sp500_df_v2.0.pkl', 'rb') as f:
    sp500_df = pickle.load(f)

removed_tickers_list  = [index for index in sp500_df.index if index not in current_sp500_df.index.to_list()]

%store removed_tickers_list

removed_tickers_df = sp500_df.loc[removed_tickers_list]

added_sp500_financial_data = s.download_financial_data_from_SEC(removed_tickers_list, removed_tickers_df, data_category)

# notice only I/S based metrics need to be updated.
updated_data_category = [
                 'EarningsPerShareDiluted',
                 'CommonStockDividendsPerShareDeclared', 
                 'NetIncomeLoss'
                 ]
added_sp500_financial_data_updated = s.convert_annual_to_quarter(added_sp500_financial_data, updated_data_category)

with open(data_folder_generate + 'removed_sp500_financial_data_v1.0.pkl', 'wb') as f:
    pickle.dump(added_sp500_financial_data_updated, f)
