==> P4 - calc_PB_factors.ipynb <==
# P4 - Calculate Price-to-Book-Value Ratio

## Overview:

This project aims to calculate price-to-book-value for the s&p500 constituents. Price/Book Value is defined as Month-end price divided by latest reported book value per share.
Techniques used in the project:

1. Processing data exceptions and outliers.  
2. Working with missing data. 

import yfinance as yf
import pandas as pd
import requests
import json
import numpy as np
import pickle
import copy
import datetime as dt
import os

data_folder_download = '../datasets/download/'
data_folder_generate = '../datasets/generate/'

%store -r removed_tickers_list
%store -r sp500_tickers

## Step 1: Read in all the necessary data files 

with open(data_folder_generate + 'sp500_financial_data_v2.0.pkl', 'rb') as f: 
    current_sp500_financial_data = pickle.load(f)

with open(data_folder_generate + 'sp500_df_v2.0.pkl', 'rb') as f: 
    sp500_df = pickle.load(f)

with open(data_folder_generate + 'removed_sp500_financial_data_v2.1.pkl', 'rb') as f: 
    removed_sp500_financial_data = pickle.load(f)

with open(data_folder_generate + 'sp500_df_v1.0.pkl', 'rb') as f: 
    current_sp500_df = pickle.load(f)

with open(data_folder_generate + 'price_df.pkl', 'rb') as f: 
    price_df = pickle.load(f)

## Step 2: Calculate book_value_per_share

def calculate_bvps(sp500_financial_data_BS):

    book_value_per_share_dict = {}  # Collect results in a dictionary first
    
    for ticker in sp500_financial_data_BS.keys():
        equity_non_null_count = sp500_financial_data_BS[ticker]['StockholdersEquity']['val'].notnull().sum()
        equity_noncontrolling_non_null_count = sp500_financial_data_BS[ticker]['StockholdersEquityIncludingPortionAttributableToNoncontrollingInterest']['val'].notnull().sum()
        
        if equity_non_null_count <= equity_noncontrolling_non_null_count:
            if sp500_financial_data_BS[ticker]['StockholdersEquityIncludingPortionAttributableToNoncontrollingInterest'].empty:
                if sp500_financial_data_BS[ticker]['Liabilities'].empty:
                    book_value = (
                    sp500_financial_data_BS[ticker]['LiabilitiesAndStockholdersEquity']['val']
                    - sp500_financial_data_BS[ticker]['LiabilitiesCurrent']['val']
                    - sp500_financial_data_BS[ticker]['LiabilitiesNoncurrent']['val']
                    )
                else: 
                    book_value = (
                        sp500_financial_data_BS[ticker]['LiabilitiesAndStockholdersEquity']['val']
                        - sp500_financial_data_BS[ticker]['Liabilities']['val']
                    )
            else: 
                book_value = (
                sp500_financial_data_BS[ticker]['StockholdersEquityIncludingPortionAttributableToNoncontrollingInterest']['val'] 
            )
        else:
            book_value = (
                sp500_financial_data_BS[ticker]['StockholdersEquity']['val'] 
            )
        # Initialize completeness counts for each measure
        common_shares_non_null_count = sp500_financial_data_BS[ticker]['CommonStockSharesOutstanding']['val'].notnull().sum()
        diluted_shares_non_null_count = sp500_financial_data_BS[ticker]['WeightedAverageNumberOfDilutedSharesOutstanding']['val'].notnull().sum()
        basic_shares_non_null_count = sp500_financial_data_BS[ticker]['WeightedAverageNumberOfSharesOutstandingBasic']['val'].notnull().sum()
    
        # Determine the measure with the most non-null values
        if diluted_shares_non_null_count >= max(common_shares_non_null_count, basic_shares_non_null_count):
            # Use 'WeightedAverageNumberOfDilutedSharesOutstanding' if it is the most complete
            book_value_per_share = book_value.loc['2013-03-31': ].divide(
                sp500_financial_data_BS[ticker]['WeightedAverageNumberOfDilutedSharesOutstanding'].loc['2013-03-31': , 'val']
            )
            
        elif basic_shares_non_null_count > max(common_shares_non_null_count, diluted_shares_non_null_count):
        # Use 'WeightedAverageNumberOfSharesOutstandingBasic' if it is the most complete
            book_value_per_share = book_value.loc['2013-03-31': ].divide(
            sp500_financial_data_BS[ticker]['WeightedAverageNumberOfSharesOutstandingBasic'].loc['2013-03-31': , 'val']
        )
        
        elif common_shares_non_null_count > max(diluted_shares_non_null_count, basic_shares_non_null_count):
            # Use 'CommonStockSharesOutstanding' if it is the most complete
            book_value_per_share = book_value.loc['2013-03-31': ].divide(
                sp500_financial_data_BS[ticker]['CommonStockSharesOutstanding'].loc['2013-03-31': , 'val']
            )
    
        # Store result in a dictionary
        book_value_per_share_dict[ticker] = book_value_per_share

    return book_value_per_share_dict

current_book_value_per_share_dict = calculate_bvps(current_sp500_financial_data)

check_result = check_accuracy(removed_book_value_per_share_dict, removed_tickers_list)

# check if the accuracy of the calculation.
def check_accuracy(book_value_per_share_dict, sp500_tickers): 
    check_result = {}
    for ticker in sp500_tickers: 
        count = book_value_per_share_dict[ticker].count()
        check_result[ticker] = count
    return check_result

sorted_dict = dict(sorted(check_result.items(), key=lambda item: item[1]))
sorted_dict

## Step 3: Uncover the reasons for the outliers. 
'A': 47,
 'ADI': 47,
 'AMAT': 47,
 'ADSK': 47,
 'BBWI': 47,
 'BF.B': 47,
 'COO': 47,
 'DG': 47,
 'HD': 47,
 'INTU': 47,
 'LDOS': 47,
 'LULU': 47,
 'NTAP': 47,
 'NDSN': 47,
 'NVDA': 47,
 'SJM': 47,
 'ULTA': 47,
 'WMT': 47,
 'JCI': 48,
 'KDP': 48,
 'DE': 49,
 'GPN': 50

#### Step 3.1 Handling exceptions: 


# current_sp500_financial_data['DE']['WeightedAverageNumberOfDilutedSharesOutstanding'].drop(['2015-11-01','2016-10-30', '2016-05-01'], inplace = True)
# current_sp500_financial_data['DE']['StockholdersEquity'].drop(['2016-05-01','2016-10-30'], inplace = True)
# current_sp500_financial_data['DE']['StockholdersEquityIncludingPortionAttributableToNoncontrollingInterest'].drop(['2014-11-02', '2015-11-01', '2016-05-01', '2016-10-30'], inplace = True)

# current_sp500_financial_data['GPN']['WeightedAverageNumberOfDilutedSharesOutstanding'].drop(['2016-09-30','2016-03-31', '2016-06-30', '2016-12-31'], inplace = True)
# current_sp500_financial_data['GPN']['StockholdersEquity'].drop(['2016-11-30'], inplace = True)
# current_sp500_financial_data['GPN']['StockholdersEquityIncludingPortionAttributableToNoncontrollingInterest'].drop(['2015-12-31', '2016-03-31', '2016-06-30', '2016-09-30', '2016-12-31'], inplace = True)

# current_sp500_financial_data['KDP']['WeightedAverageNumberOfDilutedSharesOutstanding'].drop(['2016-03-02','2016-09-24', '2016-12-24'], inplace = True)
# current_sp500_financial_data['KDP']['StockholdersEquity'].drop(['2015-09-26', '2016-03-02', '2016-09-24'], inplace = True)
# current_sp500_financial_data['JCI']['WeightedAverageNumberOfDilutedSharesOutstanding'].drop(['2016-03-25','2016-06-24'], inplace = True)
# current_sp500_financial_data['JCI']['StockholdersEquity'].drop(['2013-09-30', '2014-09-30', '2015-09-30'], inplace = True)

# removed_sp500_financial_data['ATI']['WeightedAverageNumberOfDilutedSharesOutstanding'].drop(['2022-01-02', '2023-01-01', '2023-04-02', '2023-07-02'], inplace = True)
# removed_sp500_financial_data['ATI']['StockholdersEquity'].drop(['2023-01-01'], inplace = True)
# removed_sp500_financial_data['ATI']['StockholdersEquityIncludingPortionAttributableToNoncontrollingInterest'].drop(['2021-01-03', '2022-01-02', '2023-01-01', '2023-04-02', '2023-07-02'], inplace = True)

# removed_sp500_financial_data['HRB']['WeightedAverageNumberOfDilutedSharesOutstanding'].drop(['2020-06-30', '2020-09-30', '2020-12-31', '2021-03-31', '2021-12-31'], inplace = True)
# removed_sp500_financial_data['HRB']['StockholdersEquity'].drop(['2020-05-01', '2020-06-30', '2020-09-30', '2020-12-31', '2021-03-31', '2021-05-01', '2021-12-31'], inplace = True)

# removed_sp500_financial_data['DO']['StockholdersEquity'].drop(['2021-04-23', '2021-04-24'], inplace = True)
# removed_sp500_financial_data['CHK']['WeightedAverageNumberOfDilutedSharesOutstanding'].drop(['2021-02-09'], inplace = True)
# removed_sp500_financial_data['CHK']['StockholdersEquity'].drop(['2021-02-09'], inplace = True)
#No shares outstanding data available on SEC: STZ and V

with open(data_folder_generate + 'sp500_financial_data_v2.1.pkl', 'wb') as f: 
    pickle.dump(current_sp500_financial_data, f)

with open(data_folder_generate + 'removed_sp500_financial_data_v2.1.pkl', 'wb') as f: 
    pickle.dump(removed_sp500_financial_data, f)

## Step 4: Analyze financial data set available among the S&P 500 tickers on SEC

# Identify the relevant financial labels in the database. 
item_list = {}
for ticker in sp500_tickers: 
    abt_fact= get_facts(ticker)
    item_list[ticker] = [item for item in abt_fact['facts']['us-gaap'].keys() if 'liabilities' in item.lower()]

# count the most complete financial data label
from collections import Counter

# Flatten the list of items across all tickers into a single list
all_items = [item for sublist in item_list.values() for item in sublist]

# Count the occurrences of each item
item_counts = Counter(all_items)

# Sort and rank items by most common
most_common_items = item_counts.most_common()

# Display the ranked items
for item, count in most_common_items:
    print(f"{item}: {count}")

length_list = {}
for key in book_value_per_share_dict.keys(): 
    length_list[key] = len(book_value_per_share_dict[key].values)

for key in book_value_per_share_dict.keys(): 
    book_value_per_share_dict[key] = book_value_per_share_dict[key].dropna(axis = 0)

### Availability Summary of the financial data on SEC

for shareholder's equity: 
- LiabilitiesAndStockholdersEquity: 503
- StockholdersEquity: 495
- StockholdersEquityIncludingPortionAttributableToNoncontrollingInterest: 402

for earnings per share: 
- EarningsPerShareDiluted: 498 
- EarningsPerShareBasic: 498

for shares_outstanding data: 
- CommonStockSharesOutstanding: 397
- WeightedAverageNumberOfDilutedSharesOutstanding: 498
- WeightedAverageNumberOfSharesOutstandingBasic: 497

for assets and liabilities: 
- LiabilitiesAndStockholdersEquity: 503
- LiabilitiesCurrent: 424
- Liabilities: 367
- LiabilitiesNoncurrent: 897

## Step 5: Combining current and removed tickers bvps results

len(current_book_value_per_share_dict)

len(removed_book_value_per_share_dict)

# Create combined_book_value_per_share_dict
combined_bvps_dict = {**current_book_value_per_share_dict, **removed_book_value_per_share_dict}

combined_bvps_dict.keys()

## Step 6: Calculate Price-to-Book-Value

### Step 6.1: Create monthly price table for all the tickers

price = {}
folder = '../datasets/download/companyprice/'
path = [f for f in os.listdir(folder) if not f.startswith('.')]  # Exclude hidden files
for filename in path: 
    ticker = filename[:-4]
    df = pd.read_csv(folder + filename, index_col = 'Date', parse_dates = True)
    df.index = pd.to_datetime(df.index)
    price[ticker] = df.resample('M').last()

price_df = pd.concat(price.values(), axis = 1, keys = price.keys())
price_df.columns = price_df.columns.droplevel(1)
price_df = price_df.iloc[2: ]

price_df.columns = price_df.columns.str.replace('-', '.')

price_df.dropna(how = 'all', axis = 1, inplace = True)

with open(data_folder_generate + 'price_df.pkl', 'wb') as f:
    pickle.dump(price_df, f)

price_df.head()

### Step 6.2: Forward Fill the bvps file on a monthly basis.

for ticker in combined_bvps_dict.keys():
    # Forward fill quarterly data to align with monthly price data
    combined_bvps_dict[ticker] = combined_bvps_dict[ticker].resample('M').ffill()

combined_bvps_df = pd.concat(combined_bvps_dict.values(), axis = 1, keys = combined_bvps_dict.keys())

combined_bvps_df.head()

pd.reset_option('display.max_rows')

# Count the number of NaNs in each column
na_count = combined_bvps_df.iloc[-4:].isna().sum()

# Select columns where the number of NaNs is fewer than 4
cols_to_fill = na_count[na_count < 4].index

# Apply forward fill to these columns
combined_bvps_df[cols_to_fill] = combined_bvps_df[cols_to_fill].ffill()

### Step 6.3: Divide price by bvps

price_book_value_df = price_df / combined_bvps_df

price_book_value_df.head()

with open(data_folder_generate + 'price_book_value_df_v1.1.pkl', 'wb') as f:
    pickle.dump(price_book_value_df, f) # reflected on the updated weighting. 

with open(data_folder_generate + 'price_book_value_df_v1.1.pkl', 'rb') as f:
    price_book_value_df = pickle.load(f) # reflected on the updated weighting. 

### Step 6.4 Adjust for shares outstanding

sp_500_financial2 = copy.deepcopy(removed_sp500_financial_data)

for ticker in sp_500_financial2.keys():
    # First, check if the 'WeightedAverageNumberOfDilutedSharesOutstanding' column exists and is not empty
    df = sp_500_financial2[ticker].get('WeightedAverageNumberOfDilutedSharesOutstanding')
    
    if df is not None and not df.empty:
        last_value = df.iloc[-1]['val']
        sp_500_financial2[ticker]['WeightedAverageNumberOfDilutedSharesOutstanding']['val'] = last_value
    else:
        # If 'WeightedAverageNumberOfDilutedSharesOutstanding' is empty or does not exist,
        # check 'WeightedAverageNumberOfSharesOutstandingBasic'
        df = sp_500_financial2[ticker].get('WeightedAverageNumberOfSharesOutstandingBasic')
        
        if df is not None and not df.empty:
            last_value = df.iloc[-1]['val']
            sp_500_financial2[ticker]['WeightedAverageNumberOfSharesOutstandingBasic']['val'] = last_value
        else:
            # If both of the above are empty or do not exist, check 'CommonStockSharesOutstanding'
            df = sp_500_financial2[ticker].get('CommonStockSharesOutstanding')
            
            if df is not None and not df.empty:
                last_value = df.iloc[-1]['val']
                sp_500_financial2[ticker]['CommonStockSharesOutstanding']['val'] = last_value
            else:
                print(f"No relevant data for {ticker}") 

with open(data_folder_generate + 'sp500_financial_data_v2.0.pkl', 'wb') as f: 
    pickle.dump(sp_500_financial2, f)

with open(data_folder_generate + 'removed_sp500_financial_data_v2.0.pkl', 'wb') as f: 
    pickle.dump(sp_500_financial2, f)
