==> P7 - Event_Study.ipynb <==
import numpy as np
import pandas as pd
import wrds
import datetime
import pickle
import statsmodels.api as sm
import os
import json

## Define Key Functions 

def eventstudy(json_evtdata=None, evtdata = None, model='m', estwin=100, gap=50, evtwins=-10, evtwine=10, minval=70, output='df'): 

    # Set estimation, event and gap windows
    estwins = estwin + gap + np.abs(evtwins) # Estimation window start
    estwine = gap + np.abs(evtwins) + 1 # Estimation windown end
    evtwinx = estwins + 1 # this is to separate the pre-event and post-event dates within the evtwin, evt time value (0=event date, -10=window start, 10=window end)
    evtwins = np.abs(evtwins) # Convert the negative to positive 
    evtrang = evtwins + evtwine + 1 # total event window days
   
    df = conn.raw_sql(
        f"""
        SELECT 
            a.*,
            x.*,
            c.date as rdate, 
            c.ret as ret1, 
            f.mktrf,
            f.rf,
            (f.mktrf + f.rf) as mkt,
            f.smb, 
            f.hml, 
            f.umd,
            (1 + c.ret)*(coalesce(d.dlret, 0.00) +1) - 1 as ret, 
            (1 + c.ret)*(coalesce(d.dlret, 0.00) + 1)-1-(f.mktrf + f.rf) as exret_over_mkt,
            CASE WHEN c.date BETWEEN a.estwin1 AND a.estwin2 THEN 1 ELSE 0 END AS isest,
            CASE WHEN c.date BETWEEN a.evtwin1 AND a.evtwin2 THEN 1 ELSE 0 END AS isevt,
            CASE 
                WHEN c.date BETWEEN a.evtwin1 AND a.evtwin2 
                    THEN (RANK() OVER (PARTITION BY x.evtid ORDER BY c.date) - {evtwinx}) 
                    ELSE (RANK() OVER (PARTITION BY x.evtid ORDER BY c.date))
                END AS evttime, 
            CASE WHEN c.date = a.date THEN 1 ELSE 0 END AS evtflag
            
        FROM 
        (
            SELECT 
                date,
                lag(date, {estwins}) over (order by date) as estwin1, 
                lag(date, {estwine}) over (order by date) as estwin2, 
                lag(date, {evtwins}) over (order by date) as evtwin1, 
                lead(date, {evtwine}) over (order by date) as evtwin2
            FROM crsp_a_stock.dsi
        ) as a
        JOIN 
        (
            SELECT 
                to_char(x.edate::date, 'ddMONYYYY') || trim(to_char(x.permno, '9999999999')) AS evtid,
                x.permno,
                x.edate::date
            FROM 
                json_to_recordset('{json_evtdata}'::json) AS x(edate text, permno int)
        ) AS x
            ON a.date = x.edate
        JOIN 
            crsp_a_stock.dsf c
            ON x.permno = c.permno
            AND c.date BETWEEN a.estwin1 and a.evtwin2
        JOIN ff_all.factors_daily f
            ON c.date = f.date
        LEFT JOIN crsp_a_stock.dsedelist d
            ON x.permno = d.permno
            AND c.date = d.dlstdt
        WHERE f.mktrf IS NOT NULL
        AND c.ret IS NOT NULL
        ORDER BY x.evtid, x.permno, a.date, c.date    
        """
)

    # for each event, calcualte abnormal return based on chosen risk model. 

    for evt in evtdata:
        permno = pd.to_numeric(evt['permno'])
        xdate = evt['edate']
        edate = pd.to_datetime(xdate).date()
        df['edate'] = pd.to_datetime(df['edate']).dt.date
        
        ths_mask = (df['permno'] == permno) & (df['edate'] == edate)
        est_mask = (df['permno'] == permno) & (df['edate'] == edate) & (df['isest'] == 1)
        evt_mask = (df['permno'] == permno) & (df['edate'] == edate) & (df['isevt'] == 1)
        flg_mask = (df['permno'] == permno) & (df['edate'] == edate) & (df['evtflag'] == 1)   
    
    
        #########################################################
        # check to see if it meets the min obs for estimation window
        #########################################################
    
        # number of observation in the estimation window:
        _nobs_est = df[est_mask]['ret'].count()   
        # number of actual event date flags - 
        _flgs = df[flg_mask]['ret'].count()
        # number of obersations in the event window: 
        _wins_evt = df[evt_mask]['ret'].count()
        # Only carry out the analysis if the number of obsevations meets the minimum threshold
        if (_nobs_est >= minval) and (_flgs > 0) and (_wins_evt == evtrang):

         #########################################################
         # FF Regression model
         #########################################################

            # Market Model
            if model == 'm':
            
            # Set the variables from the output
                df_est = df[est_mask]
                _nobs = len(df_est[df_est.ret.notnull()])   # not null observations
    
             # Set Y to the estimation window records
                X = df_est["mktrf"]
                df_est = df_est.assign(**{'ret-rf': df_est['ret'] - df_est['rf']})
                Y= df_est['ret-rf']
    
                # Fit an OLS model with intercept on mktrf
                X = sm.add_constant(X)
                est = sm.OLS(Y, X).fit()         
    

                alpha = est.params.const
                beta_mkt = est.params.mktrf
    
                df.loc[evt_mask, 'INTERCEPT'] = alpha
                df.loc[evt_mask, 'alpha'] = alpha
                df.loc[evt_mask, 'beta_mkt'] = beta_mkt
                df.loc[evt_mask, 'RMSE'] = np.sqrt(est.mse_resid)
                df.loc[evt_mask, '_nobs'] = _nobs
                df.loc[evt_mask, 'var_estp'] = est.mse_resid
                df.loc[evt_mask, 'rsq'] = est.rsquared
               
    
                nloc = {'alpha': alpha, 'beta_mkt': beta_mkt, 'const': 0}
    
                def f_expret(row):
                    return (nloc['alpha'] + (nloc['beta_mkt'] * row['mktrf']))
                df.loc[evt_mask, 'expret'] = df[evt_mask].apply(f_expret, axis=1)
    
                def f_abret(row):
                    return (row['ret'] - row['expret'])
                df.loc[evt_mask, 'abret'] = df[evt_mask].apply(f_abret, axis=1)
    
                nloc = {'const': 0}
    
                def f_cret(row):
                    tmp = (1+nloc['const'])*(1+row['ret']) - 1
                    nloc['const'] = tmp
                    return tmp
                df.loc[evt_mask, 'cret'] = df[evt_mask].apply(f_cret, axis=1)
                df.loc[evt_mask, 'cret_edate'] = nloc['const']
    
                nloc = {'const': 0}
    
                def f_cexpret(row):
                    tmp = (1+nloc['const'])*(1+row['expret']) - 1
                    nloc['const'] = tmp
                    return tmp
                df.loc[evt_mask, 'cexpret'] = df[evt_mask].apply(f_cexpret, axis=1)
                df.loc[evt_mask, 'cexpret_edate'] = nloc['const']
    
                nloc = {'const': 0}
    
                def f_car(row):
                    tmp = (row['abret'] + nloc['const'])
                    nloc['const'] = tmp
                    return tmp
                df.loc[evt_mask, 'car'] = df[evt_mask].apply(f_car, axis=1)
                df.loc[evt_mask, 'car_edate'] = nloc['const']
    
                nloc = {'const': 0}
    
                def f_sar(row):
                    tmp = (row['abret'] / np.sqrt(row['var_estp']))
                    nloc['const'] = tmp
                    return tmp
                df.loc[evt_mask, 'sar'] = df[evt_mask].apply(f_sar, axis=1)
                df.loc[evt_mask, 'sar_edate'] = nloc['const']
    
                nloc = {'const': 0, 'evtrang': evtrang}
    
                def f_scar(row):
                    tmp = (row['car'] / np.sqrt((evtrang * row['var_estp'])))
                    nloc['const'] = tmp
                    return tmp
                df.loc[evt_mask, 'scar'] = df[evt_mask].apply(f_scar, axis=1)
                df.loc[evt_mask, 'scar_edate'] = nloc['const']
    
                nloc = {'const': 0}
    
                def f_bhar(row):
                    tmp = (row['cret'] - row['cexpret'])
                    nloc['const'] = tmp
                    return tmp
                df.loc[evt_mask, 'bhar'] = df[evt_mask].apply(f_bhar, axis=1)
                df.loc[evt_mask, 'bhar_edate'] = nloc['const']
    
            # Fama-French model: 
            elif model == "ff": 
                df_est = df[est_mask]
                X = df_est[['smb', 'hml', 'mktrf']]
                df_est = df_est.assign(**{'ret-rf': df_est['ret'] - df_est['rf']})
                Y= df_est['ret-rf']
            
                # Fit an OLS model with intercept on mktrf, smb, hml
                X = sm.add_constant(X)
                est = sm.OLS(Y,X).fit()
            
                alpha = est.params.const
                beta_mktrf = est.params.mktrf
                beta_smb = est.params.smb
                beta_hml = est.params.hml
            
                # Add the regression results to the df
                df.loc[evt_mask, 'INTERCEPT'] = alpha
                df.loc[evt_mask, 'alpha'] = alpha
                df.loc[evt_mask, 'beta_mktrf'] = beta_mktrf
                df.loc[evt_mask, 'beta_smb'] = beta_smb
                df.loc[evt_mask, 'beta_hml'] = beta_hml
                df.loc[evt_mask, 'RMSE'] = np.sqrt(est.mse_resid)
                df.loc[evt_mask, '_nobs'] = _nobs_est
                df.loc[evt_mask, 'var_estp'] = est.mse_resid
                df.loc[evt_mask, 'rsq'] = est.rsquared
                df.loc[evt_mask, '_p_'] = 2
                df.loc[evt_mask, '_edf_'] = len(Y) - 2
                
            
                params = {'alpha': alpha, 'beta_mktrf': beta_mktrf, 'beta_smb': beta_smb, 'beta_hml': beta_hml}
                # calculate expected returns based on the Famma-French Factors
                def f_expret(row): 
                    return params['alpha'] + (params['beta_mktrf'] * row['mktrf']) + (params['beta_smb'] * row['smb']) + (params['beta_hml'] * row['hml'])+ row['rf']
                df.loc[evt_mask, 'expret'] = df[evt_mask].apply(f_expret, axis = 1)
                
                # calcualte abnormal return using the actual return minus the expected return
                def f_abret(row):
                    return (row['ret'] - row['expret'])
                df.loc[evt_mask, 'abret'] = df[evt_mask].apply(f_abret, axis = 1)
                
                
                nloc = {'const': 0}
                # calculate cumulative returns during the evtwin
                def f_cret(row):
                    tmp = (1+nloc['const'])*(1+row['ret']) - 1
                    nloc['const'] = tmp
                    return tmp
                df.loc[evt_mask, 'cret'] = df[evt_mask].apply(f_cret, axis = 1)
                df.loc[evt_mask, 'cret_edate'] = nloc['const']
                
                
                nloc = {'const': 0}
                # calculate cumulative expected returns
                def f_cexpret(row):
                    tmp = (1+nloc['const'])*(1+row['expret']) - 1
                    nloc['const'] = tmp
                    return tmp
                df.loc[evt_mask, 'cexpret'] = df[evt_mask].apply(f_cexpret, axis=1)
                df.loc[evt_mask, 'cexpret_edate'] = nloc['const']
                
                
                nloc = {'const': 0}
                # calculate cumulative abnormal returns
                def f_car(row): 
                    tmp = (row['abret'] + nloc['const'])
                    nloc['const'] = tmp
                    return tmp
                df.loc[evt_mask, 'car'] = df[evt_mask].apply(f_car, axis=1)
                df.loc[evt_mask, 'car_edate'] = nloc['const']
                
                
                nloc = {'const': 0}
                # calculate standardized abnormal returns: The standardization of ARs dampens the effect of stocks with high variances on further statistical tests
                def f_sar(row):
                    tmp = (row['abret'] / np.sqrt(row['var_estp']))
                    nloc['const'] = tmp
                    return tmp
                df.loc[evt_mask, 'sar'] = df[evt_mask].apply(f_sar, axis=1)
                df.loc[evt_mask, 'sar_edate'] = nloc['const']
                
                
                nloc = {'const': 0, 'evtrang': evtrang}
                # calculate standardized cumulative abnormal returns
                def f_scar(row):
                    tmp = (row['car'] / np.sqrt((evtrang * row['var_estp'])))
                    nloc['const'] = tmp
                    return tmp
                df.loc[evt_mask, 'scar'] = df[evt_mask].apply(f_scar, axis=1)
                df.loc[evt_mask, 'scar_edate'] = nloc['const']
                
                
                nloc = {'const': 0}
                # calculate buy and hold abnormal returns:   
                def f_bhar(row):
                    tmp = (row['cret'] - row['cexpret'])
                    nloc['const'] = tmp
                    return tmp
                df.loc[evt_mask, 'bhar'] = df[evt_mask].apply(f_bhar, axis=1)
                df.loc[evt_mask, 'bhar_edate'] = nloc['const']
                
                # df.loc[evt_mask, 'pat_scale'] = (_nobs_est - 2.00) / (_nobs_est - 4.00)
                # df.loc[evt_mask, 'pat_scale_edate'] = (_nobs_est - 2.00) / (_nobs_est - 4.00)
                
    #################################
    #  STEP 4 - OUTPUT THE RESULTS  #
    #################################

    df_sta = df[df['isevt'] == 1]
    levt = df_sta['evttime'].unique()

    columns = [
        'evttime', 
        'ret_m',
        'expret_m',
        'abret_m',
        'cret_m',
        'cexpret_m',
        'abret_t',
        'car_m',
        'scar_m',
        'cexpret_edate_m',
        'cret_edate_m',
        'car_edate_m',
        'car_edate_t',
        'sar_t',
        'scar_edate_t',
        'bhar_edate_m']

    idxlist = list(levt)
    df_stats = pd.DataFrame(index=idxlist, columns=columns)
    # df_stats = df_stats.fillna(0.00000000).infer_objects(copy=False)  # with 0s rather than NaNs

    # Event
    df_stats['evttime'] = df_sta.groupby(['evttime'])['evttime'].unique()
    # Means
    df_stats['ret_m'] = df_sta.groupby(['evttime'])['ret'].mean()
    df_stats['bhar_m'] = df_sta.groupby(['evttime'])['bhar'].mean()
    df_stats['expret_m'] = df_sta.groupby(['evttime'])['expret'].mean()
    df_stats['abret_m'] = df_sta.groupby(['evttime'])['abret'].mean()
    df_stats['cret_m'] = df_sta.groupby(['evttime'])['cret'].mean()
    df_stats['cexpret_m'] = df_sta.groupby(['evttime'])['cexpret'].mean()
    df_stats['car_m'] = df_sta.groupby(['evttime'])['car'].mean()
    df_stats['cret_edate_m'] = df_sta.groupby(['evttime'])['cret_edate'].mean()
    df_stats['cexpret_edate_m'] = df_sta.groupby(['evttime'])['cexpret_edate'].mean()
    df_stats['car_edate_m'] = df_sta.groupby(['evttime'])['car_edate'].mean()
    df_stats['bhar_edate_m'] = df_sta.groupby(['evttime'])['bhar_edate'].mean()
    df_stats['sar_m'] = df_sta.groupby(['evttime'])['sar'].mean()
    df_stats['scar_m'] = df_sta.groupby(['evttime'])['scar'].mean()
    df_stats['scar_edate_m'] = df_sta.groupby(['evttime'])['scar_edate'].mean()
    
    
    # Standard deviations
    
    df_stats['abret_v'] = df_sta.groupby(['evttime'])['abret'].std()
    df_stats['sar_v'] = df_sta.groupby(['evttime'])['sar'].std()
    df_stats['car_v'] = df_sta.groupby(['evttime'])['car'].std()
    df_stats['scar_v'] = df_sta.groupby(['evttime'])['scar'].std()
    df_stats['car_edate_v'] = df_sta.groupby(['evttime'])['car_edate'].std()
    df_stats['scar_edate_v'] = df_sta.groupby(['evttime'])['scar_edate'].std()
    
    
    # Counts   
    df_stats['n'] = df_sta.groupby(['evttime'])['evttime'].count()

    # T statistics 1
    def tstat(row, m, v, n):
        return row[m] / (row[v] / np.sqrt(row[n]))

    df_stats['abret_t'] = df_stats.apply(tstat, axis=1, args=('abret_m', 'abret_v', 'n'))
    df_stats['sar_t'] = df_stats.apply(tstat, axis=1, args=('sar_m', 'sar_v', 'n'))
    df_stats['car_edate_t'] = df_stats.apply(tstat, axis=1, args=('car_edate_m', 'car_edate_v', 'n'))
    df_stats['scar_edate_t'] = df_stats.apply(tstat, axis=1, args=('scar_edate_m', 'scar_edate_v', 'n'))
    

    # Output file: 1. Last_Event_Day, 2. Event_Window_Stats

    ## Last Event Day : 
    if model == 'm':
        maxv = max(levt)
        df_evtd = df.loc[(df['isevt'] == 1) & (df['evttime'] == maxv), ['permno', 'edate', 'cret', 'cexpret', 'alpha', 'beta_mkt', 'car', 'bhar', 'scar']]
        df_evtd.sort_values(['permno', 'edate'], ascending=[True, True], inplace = True)

    elif model == 'ff': 
        maxv = max(levt)
        df_evtd = df.loc[(df['isevt'] == 1) & (df['evttime'] == maxv), ['permno', 'edate', 'cret', 'cexpret', 'alpha', 'beta_mktrf', 'beta_hml', 'beta_smb', 'car', 'bhar', 'scar']]
        df_evtd.sort_values(['permno', 'edate'], ascending=[True, True], inplace = True)

    ## Event Window:
    df_evtwin = df_stats[['evttime', 'ret_m', 'abret_m', 'sar_m', 'cret_m', 'bhar_m', 'car_m', 'scar_m', 'n', 'sar_t', 'abret_t']]

    if output == 'df': 
        retval = {}
        retval['evt_win_stats'] = df_evtwin
        retval['last_evt_day'] = df_evtd
        return retval

def get_evtdata(df, sector, evtdate):
    
    df = df[df['GICS Sector'] == sector].copy()
    df = df.assign(**{'edate': evtdate})
     
    data = df[['permno', 'edate']]
    evtdata = [
        {
            "permno": int(row['permno']),
            "edate": row['edate']
        } for _, row in data.iterrows()
    ]
    return evtdata

## Code Execution

# read in the sp500 files with GICS industry classification.
data = pd.read_csv('merged_sp500_df.csv')

# get energy companies within the sp500
evtdata_energy = get_evtdata(data, 'Energy', '11/09/2016')
json_evtdata_energy = json.dumps(evtdata_energy).replace("'", "''")
# get financial companies within the sp500
evtdata_financials = get_evtdata(data, 'Financials', '11/09/2016')
json_evtdata_financials = json.dumps(evtdata_financials).replace("'", "''")
# get technology companies within the sp500
evtdata_technology = get_evtdata(data, 'Information Technology', '11/09/2016')
json_evtdata_technology = json.dumps(evtdata_technology).replace("'", "''")

data['GICS Sector'].unique()

# get results for each sector
## energy
ff = {}
m = {}

ff['energy'] = eventstudy(json_evtdata=json_evtdata_energy, evtdata = evtdata_energy, model='ff', estwin=100, gap=50, evtwins=-10, evtwine=10, minval=70, output='df')
ff['financials'] = eventstudy(json_evtdata=json_evtdata_financials, evtdata = evtdata_financials, model='ff', estwin=100, gap=50, evtwins=-10, evtwine=10, minval=70, output='df')
ff['technology'] = eventstudy(json_evtdata=json_evtdata_technology, evtdata = evtdata_technology, model='ff', estwin=100, gap=50, evtwins=-10, evtwine=10, minval=70, output='df')

m['energy'] = eventstudy(json_evtdata=json_evtdata_energy, evtdata = evtdata_energy, model='m', estwin=100, gap=50, evtwins=-10, evtwine=10, minval=70, output='df')
m['financials'] = eventstudy(json_evtdata=json_evtdata_financials, evtdata = evtdata_financials, model='m', estwin=100, gap=50, evtwins=-10, evtwine=10, minval=70, output='df')
m['technology'] = eventstudy(json_evtdata=json_evtdata_technology, evtdata = evtdata_technology, model='m', estwin=100, gap=50, evtwins=-10, evtwine=10, minval=70, output='df')

# combine all the model results into a summary dictionary
Summary_Result = {'ff_model': ff, 'm_model': m}

# aggregate the output information for last_event_day and comparing different sectors for _ff and _m models

for key in Summary_Result:
    for stats in Summary_Result[key]:
        if key == 'ff_model':
            Aggregate_Output[f'last_event_day_ff_{stats}'] = Summary_Result[key][stats]['last_evt_day'][['cret', 'cexpret', 'alpha', 'beta_mktrf', 'beta_hml',
           'beta_smb', 'car', 'bhar', 'scar']].mean()
        if key == "m_model": 
            Aggregate_Output[f'last_event_day_m_{stats}'] = Summary_Result[key][stats]['last_evt_day'][['cret', 'cexpret', 'alpha', 'beta_mkt', 'car', 'bhar', 'scar']].mean()

## Model Output Examples

pd.DataFrame(Aggregate_Output).T

Summary_Result['ff_model']['financials']['last_evt_day']

Summary_Result['ff_model']['energy']['evt_win_stats']

Summary_Result['m_model']['financials']['evt_win_stats']

X1 = df_12476[['smb', 'hml', 'mktrf']]
df_12476['ret-rf'] = df_12476['ret'] - df_12476['rf']
Y1= df_12476['ret-rf']

# Fit an OLS model with intercept on mktrf, smb, hml
X1 = sm.add_constant(X1)
est = sm.OLS(Y1,X1).fit()
est.summary()

X = df_12476[['mktrf']]
df_12476['ret-rf'] = df_12476['ret'] - df_12476['rf']
Y= df_12476['ret-rf']

# Fit an OLS model with intercept on mktrf, smb, hml
X = sm.add_constant(X)
est = sm.OLS(Y,X).fit()
est.summary()

pd.set_option('display.max_columns', 1000)
